Link Update. It mainly introduces new Flink functionality for learning from CajaMar data. Many changes are introduced.

+++ core-dynamic/src/main/java/eu/amidst/dynamic/datastream/filereaders/DynamicDataInstanceImpl.java
@@ -79,12 +79,12 @@ class DynamicDataInstanceImpl implements DynamicDataInstance, Serializable {
     }
 
     @Override
-    public int getSequenceID() {
+    public long getSequenceID() {
         return sequenceID;
     }
 
     @Override
-    public int getTimeID() {
+    public long getTimeID() {
         return timeID;
     }
